Digital Version of Sequence Diagram
Modifications made to add underlying processes of the calculate method

Calculate method makes calls to operand.evaluate() methods for each operand in the parameters, the result of which is returned to the calculate method to continue its process

The final results by the end of the program are stored in the o4 and o3 operand variables, but the final interaction is performed by br_smallerThan.

Each of the variables have their result printed out, but the print is handled by the main method, and not one of the class objects. This is the reasoning in not including it in the diagram. 

The interactions between the following classes are covered: BinaryOperator (Extends Operator, but overrides the calcualte method), Operand (Extends expression, but overrides the evaluate method)